Refactorizacion de tablas para mostrar info, componente Tabla creado para ser renderizado por cualquier componente, arreglos menores al codigo.
