+++ src/App.js
@@ -1,7 +1,7 @@
 import {useState} from "react";
-import { useDispatch, useSelector } from 'react-redux'
+// import { useDispatch, useSelector } from 'react-redux'
 import {BrowserRouter as Router, Switch, Route} from "react-router-dom";
-import { setLogged } from './actions/user'
+// import { setLogged } from './actions/user'
 import { ThemeProvider } from '@mui/material/styles';
 import { theme } from './theme'
 import CssBaseline from '@mui/material/CssBaseline'
