+++ src/components/container/Container.jsx
@@ -6,7 +6,7 @@ import {ContainerDiv} from './ContainerStyles'
 import Hello from '../Hello';
 import {setLogged} from "../../actions/user";
 import ProfileIndex from "../profile/ProfileIndex";
-import CreateUser from "../createUser/CreateUser"
+import UserIndex from "../User/UserIndex"
 
 const Content = ({show}) => {
   const dispatch = useDispatch()
@@ -17,7 +17,7 @@ const Content = ({show}) => {
       <section className={show && 'close'}>
         <Switch>
           <Route exact path="/roles" component={ProfileIndex} />
-          <Route exact path="/create" component={CreateUser}/>
+          <Route exact path="/users" component={UserIndex}/>
         </Switch>
        {/*  <button onClick={() => dispatch(setLogged())}>Change Logged</button>
         <p>Is logged? {isLogged.toString()}</p>
